Refactor all kinds of paginators to one
