+++ src/graphql/get-member-claims.graphql
@@ -3,6 +3,9 @@ query GetMemberClaims($memberId: ID!) {
     memberId
     claims {
       id
+      member {
+        memberId
+      }
       registrationDate
       type {
         __typename
